perf(benchmark): split dispatcher store growth baselines
