+++ wow-benchmarks/src/jmh/kotlin/me/ahoo/wow/redis/RedisCommandDispatcherBenchmark.kt
@@ -13,7 +13,10 @@
 
 package me.ahoo.wow.redis
 
+import me.ahoo.wow.api.command.CommandMessage
+import me.ahoo.wow.command.createCommandMessageForNewAggregate
 import me.ahoo.wow.eventsourcing.EventStore
+import me.ahoo.wow.example.api.cart.AddCartItem
 import me.ahoo.wow.modeling.AbstractCommandDispatcherBenchmark
 import me.ahoo.wow.redis.eventsourcing.RedisEventStore
 import org.openjdk.jmh.annotations.Benchmark
@@ -43,8 +46,12 @@ open class RedisCommandDispatcherBenchmark : AbstractCommandDispatcherBenchmark(
         return RedisEventStore(redis.redisTemplate)
     }
 
+    override fun createBenchmarkCommandMessage(): CommandMessage<AddCartItem> {
+        return createCommandMessageForNewAggregate()
+    }
+
     @Benchmark
-    override fun sendAndWaitForProcessed(blackHole: Blackhole) {
+    fun sendAndWaitForProcessedForNewAggregate(blackHole: Blackhole) {
         super.sendAndWaitForProcessed(blackHole)
     }
 }
